requirements.txt
-e git+git@github.com:tesh11/backbone-tastypie.git@cc57de49f8281354fcd9fd55402c51bd4eb2277e#egg=Backbone_tastypie-dev
Django==1.4
South==0.7.5
boto==2.4.1
distribute==0.6.24
django-storages==1.1.4
django-tastypie==0.9.11
gunicorn==0.14.2
gevent=0.13.7
mimeparse==0.1.3
psycopg2==2.4.5
python-dateutil==1.5
wsgiref==0.1.2